refactor(auth): JWT Helper class 내부 static DTO를 외부 DTO 클래스로 분리, JWT Helper class 분리된 알고리즘 참조 적용
rel: #77
